73 and 85
 "# read the data and take a look at the data and split it into train and test \n",
    "# I copied the code from chapter (because of laziness :- )\n",
    "Hitters = pd.read_csv('data/Hitters.csv', header=0, na_values='NA')\n",
    "Hitters = Hitters.dropna().reset_index(drop=True) # drop the observation with NA values and reindex the obs from 0\n",
    "dummies = pd.get_dummies(Hitters[['League', 'Division', 'NewLeague']])\n",
    "\n",
    "y = Hitters.Salary  # the response variable \n",
    "X_prep = Hitters.drop (['Salary', 'League', 'Division', 'NewLeague'], axis = 1).astype('float64')\n",
    "X = pd.concat([X_prep,  dummies[['League_A', 'Division_E', 'NewLeague_A']]], axis=1)\n",
    "X_train, X_test , y_train, y_test = train_test_split(X, y, test_size=0.66)\n"
   ]
  },
This code performs a machine learning process by loading a data set, examining the data, and then splitting it into training and test sets.

First, data is read from the 'Hitters.csv' file. The header of the read data is on line 0, NA values are left blank and line numbers are reset.

Next, dummy variables are created for the 'League', 'Division' and 'NewLeague' attributes. Dummy variables are a frequently used coding technique, especially for categorical data.

The target variable is set to 'Salary' and the data is converted to float64 data type with the "astype" method.

Finally, using the "train_test_split" method, the data is split into training and test sets. The "test_size" parameter determines the percentage of data to be kept in the test set. In this example, 66% of the data is kept in the test set and the remaining 33% is used in the training set. The training and test sets are stored in the variables "X_train", "X_test", "y_train" and "y_test" respectively.